requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
mysqlclient==2.2.0
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Twilio Integration
twilio==8.10.0

# HTTP and WebSocket
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
